Character saves needed to be async, too slow to run synchronous in client process
Fixes #44

Database class now has the ability to asynchronously run queries that don't tie up the process threads.

This is being utilized right now for character saves as that seems to be the slowest process.  However query id's (and potentially extension of such) can allow anything to be asynchronous -- and later on even add return vars.