Resolves proper rotation of antonica/commons spires
Fix #81

On CheckNode we now handle a VeXformNode with a nodeFlags of 1409875968 as resetting the yaw/pitch/roll to 0 for the next mesh.  There seems to be a VeXformNode after the mesh as well doing the same thing so I consider that 'resetting' to the original yaw/pitch/roll logic.