Fix groups persisting cross zone
Fix #268 - inadvertently display works fine most of the time, except if you cross zone the group was broken and unrecoverable due to bad saving of the group_id.

I am re-using the rejoin_group_id to set that as the default DB value in case the player loses its group temporarily when we zone them.